GER-977:  Direct GitUpdate over http with in process validation.
GitUpdate can now be performed by gitms, but be triggered in process with all validation and packfile generating happening within JGit, and not called out to the rp-git-update script.  This allow us to easily extend the objects, use serialization on the return values easily etc.  Batch operations are being held off and checked in later phase: GER-1017.

Change-Id: Ifb3109a7dcf1bc32e13684cae162fd3c0c282422