Fix for GER-1021 [Huawei] Submit failed with repository partially updated when node disconnecting/connecting to the ecosystem.
There are 2 parts to this fix, this is part 1.
GER-1017 fixed the main issue here. However with the other 2 nodes down the feed back to
the gerrit UI needed some work.
When you click the +2 it will not +2 the review but there is no feedback to UI. Secondly,
if we bring up nodes 2 and 3, +2 the review and stop the replicator on nodes 2 and 3
when we click submit we just get a generic error message and it does not tell us the reason
why it failed i.e. because there is no quorum.

Also I have enabled in process PackUtils in this review.

Change-Id: Iae8b16f9bcfba207f610e8fb4ced42c2bd341592