GER-1384: Fix openPackedObject()
Provide new ways to open a packed object that does not
get locked in a CPU spin. A scan value  > 0
will retry the open object a specific number of times.
A value of -1 will simply use existing logic and
a value of 0 will use simplified logic.

Also refactor some existing code so that it is not duplicated.

Change-Id: I54a9dcb5805d1c372410e1a3050e171bc818a978