GER-1384: Fix openPackedObject()
Provide new ways to open a packed object that does not
get locked in a CPU spin. A scan value  > 0
will retry the open object a specific number of times.
A value of -1 will simply use existing logic and
a value of 0 will use simplified logic.

Also refactor some existing code so that it is not duplicated.

Change-Id: I94b70d37235b3688fcea603dbbfca44f50a49481