GER-1336: Use WD Artifactory
Use wd_central as primary source for assets.

Point to wandisco wd_central repository as the primary source of assets and plugins.  This ensures that the assets are not only quickly downloaded but also always cached on our server incase they disappear globally from the public repos used by default in this project.

Change-Id: I8b7f600a16fed73a78ca01e7c681fc80b3567e9a