GER-1628: WindowCache holding expired file descriptors
The WindowCache holds references to packfile file descriptors.
When an out of band GC is triggered by some external tool or CGit etc.
the cache has no way of knowing when to evict a stale entry.

This implementation:
* provides a scheduledable task via exposed git config properties;
* runs through each cache entry;
* calculates locks appropriately & only if necessary;
* cleans up entry resources in the same fashion as removeAll()
 methods of WindowCache.
* local copy of the WindowCache instance is created to avoid access
to the volatile member var (even though it is in scope),
testing showed a slight increase in efficiency.
* adds a test which triggers the task directly - this avoids
 Thread.sleep() in the tests.
* checks cached snapshot fileKey against filesystem filekey at runtime
to 100% check the entry is legit.

Change-Id: I8983e04ebdf9ae67471115e2e859d2daed300130