NV-8684: Support Mixed Batch Failures
Prevent retry on mixed tag creation state.

Added ability to bounce the creation of a ref which already exists but
which is mixed with other items which do not exist.  So if we aren't
all in new or all in old we deny the batch.  But a new error prevents
the auto retry being used by indicating REJECTED_OTHER_REASON
and special error text we check for in the gitms before it gets
back to the client.

Change-Id: Icacd957ce8090891a851a6ac6c5818d8f2f64531